Add VarName, VarCode to CUAHSI search results
Instead of list of concept_keyword strings, we now return a
list of objects with keys: code, name, and concept_keyword.
This will allow us to continue using concept_keyword rather
than the very verbose name as the display field, and adding
it under the hood to the code necessary for fetching detail
values.

We still serve the name, in case the full form is needed.

Some sample values:

{
    "code": "NWISUV:63680",
    "concept_keyword": "Turbidity",
    "name": "Turbidity, water, unfiltered, monochrome near
             infra-red LED light, 780-900 nm, detection angle 90
             +-2.5 degrees, formazin nephelometric units (FNU)"
}

{
    "code": "NWISUV:00301",
    "concept_keyword": "Oxygen, dissolved percent of saturation",
    "name": "Dissolved oxygen, water, unfiltered, percent of
             saturation"
}

{
    "code": "NWISUV:99988",
    "concept_keyword": "Radiation, incoming PAR",
    "name": "Photosynthetically active radiation (average flux
             density on a horizontal surface during measurement
             interval), micromoles of photons per square meter
             per second"
}